split config from datahelper
